+++ src/AuthProvider/TokenOAuthProvider.php
@@ -9,6 +9,7 @@ class TokenOAuthProvider implements AuthProviderInterface
     public const GRANT_PASSWORD = "password";
     public const GRANT_CODE     = "authorization_code";
     public const GRANT_FIXED_TOKEN = "fixed_token";
+    public const GRANT_CREDENTIALS = "client_credentials";
 
     /**
      * @var bool
@@ -93,10 +94,18 @@ class TokenOAuthProvider implements AuthProviderInterface
         ?string $password,
         ?string $accessToken,
         ?string $instanceUrl,
-        string $grantType = self::GRANT_PASSWORD,
+        ?string $grantType = null,
         ?string $redirectUri = null,
         ?string $code = null
     ) {
+        if (empty($grantType)) {
+            if (!empty($password)) {
+                $grantType = self::GRANT_PASSWORD;
+            } else {
+                $grantType = self::GRANT_CREDENTIALS;
+            }
+        }
+
         $this->clientId     = $clientId;
         $this->clientSecret = $clientSecret;
         $this->username     = $username;
@@ -169,6 +178,21 @@ class TokenOAuthProvider implements AuthProviderInterface
                         'client_secret' => $this->clientSecret,
                         'username'      => $this->username,
                         'password'      => $this->password,
+                    ],
+                    'headers'     => [
+                        'Content-Type' => 'application/x-www-form-urlencoded',
+                        'Accept'       => 'application/json',
+                    ],
+                ]
+            );
+        } elseif (self::GRANT_CREDENTIALS === $this->grantType) {
+            $response = $this->httpClient->post(
+                '/services/oauth2/token',
+                [
+                    'form_params' => [
+                        'grant_type'    => self::GRANT_CREDENTIALS,
+                        'client_id'     => $this->clientId,
+                        'client_secret' => $this->clientSecret,
                     ],
                     'headers'     => [
                         'Content-Type' => 'application/x-www-form-urlencoded',